Setup a topicmap renderer based on d3.js
Works rudimentarily.
To see a D3 based graph rendering (with forces): create a normal topicmap and then change its Topicmap Renderer URI to `dm4.d3js_renderer` manually (in the Webclient). To do so you must set the `Hidden` and `Locked` flags of the Topicmap Renderer URI's View Configutation to `false`. (`Topicmap Renderer URI` is a child type of `Topicmap`).